Move tina config
