+++ pages/[slug].tsx
@@ -1,4 +1,4 @@
-import { client } from '../.tina/__generated__/client'
+import { client } from '../tina/__generated__/client'
 import { GetStaticProps, GetStaticPaths } from 'next'
 import { fileToUrl } from 'utils/urls'
 import { useTina } from 'tinacms/dist/react'
@@ -18,7 +18,7 @@ const Page = props => {
 }
 
 // Data Fetching
-export const getStaticProps: GetStaticProps = async function({
+export const getStaticProps: GetStaticProps = async function ({
   preview,
   previewData,
   ...ctx
@@ -37,7 +37,7 @@ export const getStaticProps: GetStaticProps = async function({
   }
 }
 
-export const getStaticPaths: GetStaticPaths = async function() {
+export const getStaticPaths: GetStaticPaths = async function () {
   const pages = await fg(`./content/blocksPages/*.json`)
   const paths = pages.map(file => {
     const slug = fileToUrl(file, 'blocksPages')
